requirements.txt
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
dotenv==0.9.9
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.70.0
httplib2==0.31.0
idna==3.11
mysql==0.0.3
mysql-connector-python==9.5.0
mysqlclient==2.2.7
oauthlib==3.3.1
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.3.1
python-dotenv==1.1.1
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
uritemplate==4.2.0
urllib3==2.6.2
